+++ repro/models/lin2004.py
@@ -1,0 +1,2 @@
+from ._models.lin2004.src.model import ROUGE
+from ._models.lin2004.src.setup import Lin2004SetupSubcommand