+++ repro/models/sacrerouge.py
@@ -1,2 +1,2 @@
-from ._models.sacrerouge.src.models import ROUGE
+from ._models.sacrerouge.src.models import SRROUGE
 from ._models.sacrerouge.src.setup import SacreROUGESetupSubcommand